+++ app/assets/javascripts/application.js
@@ -1,0 +1,7 @@
+//= require jquery
+//= require jquery_ujs
+//= require turbolinks
+//= require moment 
+//= require fullcalendar
+//= require fullcalendar/locale-all
+//= require_tree ./modules
