+++ app/models/event.rb
@@ -6,11 +6,4 @@ class Event < ApplicationRecord
   validates :start_day, presence: true
   validates :start_hour, presence: true
 
-  def start_time
-    self.start_day
-  end
-
-  def end_time
-    self.end_day
-  end
 end
