+++ app/views/calendars/show.html.haml
@@ -11,29 +11,4 @@
     = render "comments/index"
 
   #main__calendar
-    -# = month_calendar events: @events do |date, events|
-    -#   .day__num
-    -#     = date.day
-    -#   - if HolidayJp.holiday?(Date.new(date.year, date.month, date.day))
-    -#   - holidays = HolidayJp.between(Date.new(date.year, date.month, date.day), Date.new(date.year, date.month, date.day))
-    -#     .holiday= holidays[0].name
-    -#   - events.each do |event|
-    -#     .event{class: "event-#{event.id}"}
-    -#       = link_to event.title, edit_calendar_event_path(event.calendar_id, event.id)
-    .monthCalendar
-      .dayOfTheWeek
-        .day-week
-          日
-        .day-week
-          月
-        .day-week
-          火
-        .day-week
-          水
-        .day-week
-          木
-        .day-week
-          金
-        .day-week
-          土
-      #weeks
+    #calendar{data: {"turbolinks":"false"}}
