+++ app/views/events/index.html.haml
@@ -1,25 +1,1 @@
--# %h1 Listing events
 
--# %table
--#   %thead
--#     %tr
--#       %th Name
--#       %th Start time
--#       %th End time
--#       %th
--#       %th
--#       %th
-
--#   %tbody
--#     - @events.each do |event|
--#       %tr
--#         %td= event.name
--#         %td= event.start_time
--#         %td= event.end_time
--#         %td= link_to 'Show', event
--#         %td= link_to 'Edit', edit_event_path(event)
--#         %td= link_to 'Destroy', event, method: :delete, data: { confirm: 'Are you sure?' }
-
--# %br
-
--# = link_to 'New Event', new_event_path
