+++ app/views/events/index.json.jbuilder
@@ -1,1 +1,7 @@
-# json.array! @events, partial: "events/event", as: :event
+json.array!(@events) do |event|
+  json.extract! event, :id, :title, :content 
+  json.start event.start_day
+  json.end event.end_day
+  json.url events_url(event, format: :html) 
+  # binding.pry
+end
