--- app/views/events/show.json.jbuilder
@@ -1,1 +1,0 @@
-# json.partial! "events/event", event: @event
