--- app/views/simple_calendar/_month_calendar.html.haml
@@ -1,38 +1,0 @@
-.simple-calendar{class: "container"}
-  .calendar__month{class: "table-responsive"}
-  .calendar__heading{class: "d-flex justify-content-between align-items-center"}
-    .calendar__heading--left{class: "d-flex"}
-      #previous{class: "ml-5"}
-        = link_to calendar.url_for_previous_view do
-          = icon('fas', 'caret-square-left')
-        %span.pop 前月
-      #next{class: "ml-3"}
-        = link_to calendar.url_for_next_view do
-          = icon('fas', 'caret-square-right')
-        %span.pop 翌月
-      .calendar__title{class: "ml-5"}
-        = start_date.year.to_s + "年"
-        = t('date.month_names')[start_date.month]
-    .calendar__headind--right{class: "d-flex"}
-    .change__contents{class: "d-flex justify-content-around"}
-      -# #today{class: "btn btn-light"}
-      -#   = link_to '今日', '#'
-      -#   %span.pop 今日のカレンダーへ
-      -# #change__year{class: "mx-3 btn btn-light"}
-      -#   年
-      -# #change__month{class: "btn btn-light"}
-      -#   月
-  %table.table
-    %thead
-      %tr
-        - date_range.slice(0, 7).each do |day|
-          %th= t('date.abbr_day_names')[day.wday]
-    %tbody.weeks
-      - date_range.each_slice(7) do |week|
-        %tr.week
-          - week.each do |day|
-            = content_tag :td, class: calendar.td_classes_for(day) do
-              - if defined?(Haml) && respond_to?(:block_is_haml?) && block_is_haml?(passed_block)
-                - capture_haml(day, sorted_events.fetch(day, []), &passed_block)
-              - else
-                - passed_block.call day, sorted_events.fetch(day, [])
