--- app/views/simple_calendar/_week_calendar.html.haml
@@ -1,24 +1,0 @@
-.simple-calendar
-  .calendar-heading
-    = link_to t('simple_calendar.previous', default: 'Previous'), calendar.url_for_previous_view
-    - if calendar.number_of_weeks == 1
-      %span.calendar-title
-        Week #{calendar.week_number}
-    - else
-      %span.calendar-title
-        Week #{calendar.week_number} - #{calendar.end_week}
-    = link_to t('simple_calendar.next', default: 'Next'), calendar.url_for_next_view
-  %table.table.table-striped
-    %thead
-      %tr
-        - date_range.slice(0, 7).each do |day|
-          %th= t('date.abbr_day_names')[day.wday]
-    %tbody
-      - date_range.each_slice(7) do |week|
-        %tr
-          - week.each do |day|
-            = content_tag :td, class: calendar.td_classes_for(day) do
-              - if defined?(Haml) && respond_to?(:block_is_haml?) && block_is_haml?(passed_block)
-                - capture_haml(day, sorted_events.fetch(day, []), &passed_block)
-              - else
-                - passed_block.call day, sorted_events.fetch(day, [])
